+++ application/metaData/projectGroup/app.go
@@ -29,7 +29,7 @@ func (app *app) ToInfo(id int) Dto {
 }
 
 // Count 项目组数量
-func (app *app) Count() int {
+func (app *app) Count() int64 {
 	return app.repository.Count()
 }
